Extract Logger definition to a helper method.
